+++ src/patch/core/Practice.zig
@@ -76,6 +76,12 @@ const mode_vis = struct {
 
 // HOOK FUNCTIONS
 
+pub fn OnInit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
+
+pub fn OnInitLate(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
+
+pub fn OnDeinit(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {}
+
 pub fn InitRaceQuadsA(_: *GlobalSt, _: *GlobalFn) callconv(.C) void {
     mode_vis.init();
 }