+++ src/patch/util/memory.zig
@@ -24,19 +24,23 @@ pub fn write(offset: usize, comptime T: type, value: T) usize {
     const addr: [*]align(1) T = @ptrFromInt(offset);
     const data: [1]T = [1]T{value};
     var protect: win.DWORD = undefined;
-    _ = win.VirtualProtect(addr, @sizeOf(T), win.PAGE_EXECUTE_READWRITE, &protect) catch @panic("failed to set PAGE_EXECUTE_READWRITE for memory write operation");
+    _ = win.VirtualProtect(addr, @sizeOf(T), win.PAGE_EXECUTE_READWRITE, &protect) catch
+        @panic("failed to set PAGE_EXECUTE_READWRITE for memory write operation");
     @memcpy(addr, &data);
-    _ = win.VirtualProtect(addr, @sizeOf(T), protect, &protect) catch @panic("failed to restore previous protection after memory write operation");
+    _ = win.VirtualProtect(addr, @sizeOf(T), protect, &protect) catch
+        @panic("failed to restore previous protection after memory write operation");
     return offset + @sizeOf(T);
 }
 
-pub fn write_bytes(offset: usize, ptr_in: ?*anyopaque, len: usize) usize {
+pub fn write_bytes(offset: usize, ptr_in: ?*const anyopaque, len: usize) usize {
     const addr: [*]align(1) u8 = @ptrFromInt(offset);
-    const data: []u8 = @as([*]u8, @ptrCast(ptr_in))[0..len];
+    const data: []const u8 = @as([*]const u8, @ptrCast(ptr_in))[0..len];
     var protect: win.DWORD = undefined;
-    _ = win.VirtualProtect(addr, len, win.PAGE_EXECUTE_READWRITE, &protect) catch @panic("failed to set PAGE_EXECUTE_READWRITE for memory write_bytes operation");
+    _ = win.VirtualProtect(addr, len, win.PAGE_EXECUTE_READWRITE, &protect) catch
+        @panic("failed to set PAGE_EXECUTE_READWRITE for memory write_bytes operation");
     @memcpy(addr, data);
-    _ = win.VirtualProtect(addr, len, protect, &protect) catch @panic("failed to restore previous protection after memory write_bytes operation");
+    _ = win.VirtualProtect(addr, len, protect, &protect) catch
+        @panic("failed to restore previous protection after memory write_bytes operation");
     return offset + len;
 }